+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/GetUpdates.kt
@@ -4,6 +4,8 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleReque
 import com.github.insanusmokrassar.TelegramBotAPI.types.ALL_UPDATES_LIST
 import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.RawUpdate
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.Update
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.UpdateSerializerWithoutDeserialization
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.internal.ArrayListSerializer
@@ -35,8 +37,10 @@ data class GetUpdates(
     val limit: Int? = null,
     val timeout: Int? = null,
     val allowed_updates: List<String>? = ALL_UPDATES_LIST
-): SimpleRequest<List<RawUpdate>> {
+): SimpleRequest<List<Update>> {
     override fun method(): String = "getUpdates"
 
-    override fun resultDeserializer(): KSerializer<List<RawUpdate>> = ArrayListSerializer(RawUpdate.serializer())
+    override fun resultDeserializer(): KSerializer<List<Update>> = ArrayListSerializer(
+        UpdateSerializerWithoutDeserialization
+    )
 }