+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/updateshandlers/KtorUpdatesPoller.kt
@@ -109,9 +109,7 @@ class KtorUpdatesPoller(
                 timeoutSeconds,
                 allowedUpdates
             )
-        ).map {
-            it.asUpdate
-        }
+        )
     }
 
     private suspend fun handleUpdates(updates: List<Update>) {